requirements.txt
Flask>=2.0
scikit-learn
pandas
mysql-connector-python
numpy